# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReplyX API",
    "description": "\nReplyX - Comprehensive AI Assistant Platform API\n\n## Features\n- 🤖 AI Assistant Management\n- 💬 Real-time Dialog System  \n- 📄 Document Processing & Knowledge Base\n- 🤖 Telegram Bot Integration\n- 💰 Balance & Billing System\n- 👥 User Management & Authentication\n- 📊 Analytics & Monitoring\n- 🛡️ Security & Rate Limiting\n\n## Authentication\nAll endpoints (except public ones) require JWT authentication:\n- Include `Authorization: Bearer <token>` header\n- Include `X-CSRF-Token` header for state-changing operations\n\n## Base URLs\n- Production: `https://api.replyx.ru`\n- Development: `http://localhost:8000`\n        ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.replyx.ru",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/admin/users/{user_id}/balance/topup": {
      "post": {
        "summary": "Adjust User Balance",
        "description": "Пополнить или списать баланс пользователя",
        "operationId": "adjust_user_balance_api_admin_users__user_id__balance_topup_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/balance/welcome-bonus-status": {
      "get": {
        "tags": [
          "balance"
        ],
        "summary": "Get Welcome Bonus Status",
        "description": "Проверить статус приветственного бонуса",
        "operationId": "get_welcome_bonus_status_api_balance_welcome_bonus_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/balance/claim-welcome-bonus": {
      "post": {
        "tags": [
          "balance"
        ],
        "summary": "Claim Welcome Bonus",
        "description": "Получить приветственный бонус (если еще не получен)",
        "operationId": "claim_welcome_bonus_api_balance_claim_welcome_bonus_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/balance/stats": {
      "get": {
        "tags": [
          "balance"
        ],
        "summary": "Get Balance Stats",
        "description": "Получить статистику баланса пользователя",
        "operationId": "get_balance_stats_api_balance_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/balance/current": {
      "get": {
        "tags": [
          "balance"
        ],
        "summary": "Get Current Balance",
        "description": "Получить текущий баланс пользователя",
        "operationId": "get_current_balance_api_balance_current_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/balance/topup": {
      "post": {
        "tags": [
          "balance"
        ],
        "summary": "Top Up Balance",
        "description": "Пополнить баланс пользователя",
        "operationId": "top_up_balance_api_balance_topup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/balance/transactions": {
      "get": {
        "tags": [
          "balance"
        ],
        "summary": "Get Transactions",
        "description": "Получить историю транзакций пользователя",
        "operationId": "get_transactions_api_balance_transactions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BalanceTransactionRead"
                  },
                  "title": "Response Get Transactions Api Balance Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/balance/transactions/detailed": {
      "get": {
        "tags": [
          "balance"
        ],
        "summary": "Get Detailed Transactions",
        "description": "Получить детализированную историю транзакций пользователя с информацией о связанных сущностях",
        "operationId": "get_detailed_transactions_api_balance_transactions_detailed_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BalanceTransactionDetailRead"
                  },
                  "title": "Response Get Detailed Transactions Api Balance Transactions Detailed Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/balance/check/{service_type}": {
      "get": {
        "tags": [
          "balance"
        ],
        "summary": "Check Balance For Service",
        "description": "Проверить достаточность средств для услуги",
        "operationId": "check_balance_for_service_api_balance_check__service_type__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Service Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/balance/usage-stats": {
      "get": {
        "tags": [
          "balance"
        ],
        "summary": "Get Usage Stats",
        "description": "Получить полную статистику использования (без ограничений по количеству)",
        "operationId": "get_usage_stats_api_balance_usage_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/login"
          }
        }
      }
    }
  }
}